Fixed assertion error.
Realized that the job length after being analyzed by GPT can be 0 technically. So instead, I changed `assert len(parsed_values) > 0` to a conditional to test the length. If the `len(parsed_values)` ends up being 0, then theres no point in retrieving those filtered jobs if it doesn't exist. Therefore, I simply added a descriptive print statement as well as invoking the `exit()` method to exit the entire function. This is so it doesn't call the next function where it returns the filtered jobs.