Minor updates for Issue #18
I have updated the URL to display only internships from LinkedIn in the US and Canada, providing us with more relevant opportunities. Fixes #18 !

As we aim to make this platform more accessible for other clubs to use, I have taken into consideration the need for a simple setup. Now, users are reminded to call the create() function before they start using the platform. This information has been added to the README.md file and the main() function in main.py has been updated accordingly.

To ensure clarity and accuracy, I have made a minor improvement to the docstring on the GitHub internship function.

Based on feedback and to maximize the available character limit for descriptions in embed messages, I have increased the limit of opportunities by 5 for both internships and jobs.