Updated .env file to accomodate for U.S. based LinkedIn internships. (#24)
* Minor updates for Issue #18

I have updated the URL to display only internships from LinkedIn in the US and Canada, providing us with more relevant opportunities. Fixes #18 !

As we aim to make this platform more accessible for other clubs to use, I have taken into consideration the need for a simple setup. Now, users are reminded to call the create() function before they start using the platform. This information has been added to the README.md file and the main() function in main.py has been updated accordingly.

To ensure clarity and accuracy, I have made a minor improvement to the docstring on the GitHub internship function.

Based on feedback and to maximize the available character limit for descriptions in embed messages, I have increased the limit of opportunities by 5 for both internships and jobs.

* Created Extra line in .env.example.

* Created MAX_LIST_LENGTH to hold a constant value for the max length of opportunities, created cli-command for users to create their table in their DB (created conditional to ensure the function gets called), returned an object in the cli function to hold both days needed command as well as create table command.

Look at title.

* Returned all arguments for cli commands, updated what argument value we need when we call the cli function, and finally updated readme to accomodate for the `--create` cli command.

* Removed unused comments from utility.py.

See title.

* Updated main.py and utility.py to update variable name holding days_needed cli command value.

* Updated LinkedIn internship link to accommodate only for U.S. based internships.

Look at title.

---------

Co-authored-by: Ethan Davidson <31261035+EthanThatOneKid@users.noreply.github.com>